refactor(mllp-client): fold connection + duplex into client.ts
Collapse the per-connection wire (createConnection: read loop, ACK exchange,
drop detection, teardown) and the MllpDuplex/MllpConnector adapter contract into
client.ts. The wire logic is moved verbatim (proven races intact) as a private
factory; the adapter types are defined and exported from client.ts.

connection.ts and duplex.ts deleted; index.ts and runtime/node.ts now import the
adapter types from ./client. 96 tests pass; @glion/cli still type-checks.

The cluster the maintainer flagged (duplex/response/send/connection/client) is
now client.ts + message.ts. Package: 12 → 7 modules.
